refactor: ShowsForYou section
